# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaperMan",
    "version": "0.1.0"
  },
  "paths": {
    "/api/catalog": {
      "get": {
        "summary": "Catalog",
        "operationId": "catalog",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalog"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/{kind}": {
      "post": {
        "summary": "Create Entry",
        "operationId": "create_entry",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "owners",
                "tags"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog/{kind}/{entry_id}": {
      "put": {
        "summary": "Update Entry",
        "operationId": "update_entry",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "owners",
                "tags"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$",
              "title": "Entry Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntryInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogEntry"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Entry",
        "operationId": "delete_entry",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "owners",
                "tags"
              ],
              "type": "string",
              "title": "Kind"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$",
              "title": "Entry Id"
            }
          },
          {
            "name": "reassign_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[a-zA-Z0-9_-]{1,80}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reassign To"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntryRemoval"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/documents": {
      "get": {
        "summary": "Documents",
        "operationId": "documents",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "delivery",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "",
                "review",
                "delivered"
              ],
              "type": "string",
              "default": "",
              "title": "Delivery"
            }
          },
          {
            "name": "inbox",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Inbox"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "date_desc",
                "date_asc",
                "title",
                "title_desc",
                "owners_asc",
                "owners_desc",
                "tags_asc",
                "tags_desc",
                "verification_asc",
                "verification_desc",
                "processed_asc",
                "processed_desc"
              ],
              "type": "string",
              "default": "date_desc",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}": {
      "get": {
        "summary": "Document",
        "operationId": "document",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Edit Document",
        "operationId": "edit_document",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$",
              "title": "Document Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentEdit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Catalog": {
        "properties": {
          "owners": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "type": "array",
            "title": "Owners"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/CatalogEntry"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "owners",
          "tags"
        ],
        "title": "Catalog"
      },
      "CatalogEntry": {
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          },
          "icon": {
            "type": "string",
            "enum": [
              "auto",
              "file",
              "receipt",
              "shield",
              "bank",
              "health",
              "utilities",
              "contract",
              "home",
              "car",
              "business",
              "education",
              "travel",
              "tax"
            ],
            "title": "Icon",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "aliases",
          "icon"
        ],
        "title": "CatalogEntry"
      },
...
      "DocumentDetail": {
        "properties": {
          "document": {
            "$ref": "#/components/schemas/Document"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "source": {
            "$ref": "#/components/schemas/SourceReference"
          },
          "can_manage_access": {
            "type": "boolean",
            "title": "Can Manage Access"
          },
          "can_edit_pages": {
            "type": "boolean",
            "title": "Can Edit Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document",
          "text",
          "source",
          "can_manage_access",
          "can_edit_pages"
        ],
        "title": "DocumentDetail"
      },
      "DocumentEdit": {
        "properties": {
          "owner_ids": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "minItems": 1,
            "title": "Owner Ids"
          },
          "source_pages": {
            "anyOf": [
              {
                "items": {
                  "type": "integer",
                  "minimum": 1.0
                },
                "type": "array",
                "minItems": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Pages"
          },
          "rotations": {
            "items": {
              "$ref": "#/components/schemas/PageRotation"
            },
            "type": "array",
            "title": "Rotations"
          },
          "revision": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Revision"
          },
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "document_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Date"
          },
          "summary": {
            "type": "string",
            "maxLength": 10000,
            "title": "Summary"
          },
          "tag_ids": {
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_-]{1,80}$"
            },
            "type": "array",
            "title": "Tag Ids"
          },
          "text": {
            "type": "string",
            "maxLength": 1000000,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "owner_ids",
          "revision",
          "title",
          "document_date",
          "summary",
          "tag_ids",
          "text"
        ],
        "title": "DocumentEdit"
      },
      "DocumentPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "pages"
        ],
        "title": "DocumentPage"
      },
...
      "EntryInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "icon": {
            "type": "string",
            "enum": [
              "auto",
              "file",
              "receipt",
              "shield",
              "bank",
              "health",
              "utilities",
              "contract",
              "home",
              "car",
              "business",
              "education",
              "travel",
              "tax"
            ],
            "title": "Icon",
            "default": "auto"
          },
          "aliases": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Aliases"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "EntryInput"
      },
      "EntryRemoval": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "removed",
              "in_use"
            ],
            "title": "Status"
          },
          "documents": {
            "type": "integer",
            "title": "Documents"
          },
          "scans": {
            "type": "integer",
            "title": "Scans"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "status",
          "documents",
          "scans"
        ],
        "title": "EntryRemoval"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PageRotation": {
        "properties": {
          "page": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Page",
            "description": "One-based source page number."
          },
          "clockwise": {
            "type": "integer",
            "enum": [
              90,
              180,
              270
            ],
            "title": "Clockwise",
            "description": "Clockwise correction needed to make the supplied page image upright."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "page",
          "clockwise"
        ],
        "title": "PageRotation"
      },
...
      "SourceReference": {
        "properties": {
          "scan_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_-]{1,80}$",
            "title": "Scan Id"
          },
          "inbox": {
            "type": "string",
            "enum": [
              "personal",
              "shared"
            ],
            "title": "Inbox"
          },
          "accessible": {
            "type": "boolean",
            "title": "Accessible"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scan_id",
          "inbox",
          "accessible"
        ],
        "title": "SourceReference"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}